+++ CHANGELOG.rst
@@ -6,6 +6,13 @@ Change Log
 ==============
 
 
+New Features
+------------
+
+* Added ``repo.size_nbytes`` and ``repo.size_human`` to report disk usage of a repository on disk.
+  (`#174 <https://github.com/tensorwerk/hangar-py/pull/174>`__) `@rlizzo <https://github.com/rlizzo>`__
+
+
 Bug Fixes
 ---------